refactoring перенесла апи и модели внутрь feature_places
