+++ lib/features/common/constants/app_default_values.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:places_elementary/features/common/domain/entity/object_position.dart';
+import 'package:places_elementary/features/feature_places/domain/entity/object_position.dart';
 
 /// Дефолтные значения
 class AppDefaultValues {
