+++ lib/features/feature_places/domain/entity/card_type.dart
