+++ lib/features/feature_places/domain/entity/object_position.dart
