+++ lib/features/feature_places/domain/entity/place_type.dart
