+++ lib/features/feature_places/domain/entity/search_filter.dart
